spread out the game.ts code into separate files
